+++ src/model/square/chance-square.ts
@@ -1,0 +1,7 @@
+import { type Command, SpecialSquare } from "./special-square";
+
+export class ChanceSquare extends SpecialSquare {
+  getCommand(): Command {
+    throw new Error("Method not implemented.");
+  }
+}
