+++ src/model/square/index.ts
@@ -1,1 +1,4 @@
+export { ChanceSquare } from "./chance-square";
+export { GoToStartSquare, MoveSquare } from "./move-square";
+export { SpecialSquare } from "./special-square";
 export { Square } from "./square";
