+++ src/model/square/special-square.ts
@@ -1,0 +1,10 @@
+import type { GameContext } from "../gameContext";
+import { Square } from "./square";
+
+export interface Command {
+  execute(context: GameContext): void;
+}
+
+export abstract class SpecialSquare extends Square {
+  abstract getCommand(): Command;
+}
