+++ tests/game-with-mocked-dice.test.ts
@@ -21,7 +21,7 @@ import type { Battle } from "@/model/battle";
 import { SquaresBuilder } from "@/model/board";
 import { Dice } from "@/model/dice";
 import { Game } from "@/model/game";
-import { MoveSquare, Square } from "@/model/square/square";
+import { MoveSquare, Square } from "@/model/square";
 
 describe("Game with mocked dice", () => {
   afterAll(() => {
